Adapte le code en fonction de l'implementation de Vector

+++ src/finit_difference/development/Solver.java
@@ -1,6 +1,6 @@
 package finit_difference.development;
 
 public interface Solver {
-    Vectorizable solve(double alpha, double beta, long n, Function f) throws FiniteDifferenceException;
+    Vectorizable solve(double alpha, double beta, int n, Function f) throws FiniteDifferenceException;
     // alpha = u(0); beta = u(1)
 }
